dont open links so aggressively
